[luci-app-pagekite] Support only pagekitec backend.
This package had rotted, not being updated from the move&split to
pagekitec and pagekitepy.  Because I can't work out how to make a
package depend on either pagekitec or pagekitepy, I've made this a hard
package for pagekitec.  This is the version that you would want on
openwrt anyway, it's much much lighter than the python version.