[mosquitto] Don't override try_private defaults
0.15 is ancient now, so forcing an override of the mosquitto default
try_private setting to work around a bug in 0.15 is not really what we
should be doing.

If someone has _explicitly_ added try_private settings in their UCI
config file, that setting will be observed, but if it was not set, we
let mosquitto choose it's own default for this setting.