mosquitto: use -rpath-link hacks for dependent libs
Some of the openwrt toolchains get upset about cascaded library
dependencies, and not listing them all at the command line.

In this case, as the mosquitto client library only lists "-lmosquitto"
but not the "-lssl -lcrypto" that libmosquitto in turn depends on, it
was complaining that they couldn't be found.

Explicitly listing -lssl and -lcrypto in the client build works, but
feels even more gross than this modification to add -rpath-link, which
is commonly done for dozens of openwrt packages.

NOTE: this does NOT seem to be needed in current openwrt trees! this
appears to be a problem only when using very old openwrt trees!