output-dexma: convert to use senml+cabinet model
We now _only_ process senml data.  All livestream data has moved to
senml. We now _only_ process the datacache if we have a cabinet model.
As a cabinet model is ~required for sane operation (it provides the
phase assignments, CT sizes and breaker sizes and names) this seems
reasonable, and it means we can create meaningful datapoint names for
the dexma parameters.

We provide the following values at this point:
402 cumulative kwh
404 cumulative kvarh
425 max current in interval
426 averge current in interval

Messages are posted every 15 minutes.

Signed-off-by: Karl Palsson <karlp@etactica.com>