output-db: store simply into a flat table
Don't try and normalize the datapoint ids.  Sure, it saves disk space,
but it means a lot more caching and looking up ids, and not a huge
amount of gain.  This was much more important when we were writing to a
local sd card, but ~nothing when we're writing to a cloud mysql/postgres
server