output-db: externalize the metadata insert query
Instead of select? insert+update, do update, then insert+update for the
"upsert" of metadata.  This means the normal case stays a single query,
and the rare case is three queries, as before.  The extra advantage is
that now there are only two queries, so just bite the bullet and make
that a parameter as well.

Move all the queries onto their own tab of "really advanced" stuff.