output-db: replace "download schema" will full editor
Just let the user edit the schema directly, instead of only allowing a
download.  We already have the mechanism in place for editing the
queries.  This drops the code for doing special custom buttons, which
might be interesting for things at a later date, and will need to be dug
up again!

As with the original download schema button, this doesn't work until
you've saved the initial new instance, so that the selected database
driver is available.