output-db: queue data readings if database is unavailable
We won't try and queue metadata, those are available whenever we
restart, and have more complicated insertions, so just abort and let
process monitoring restart.

But the interval data is "precious"  Queue it, (up to a limit) and
attempt to reconnect to the database periodically.  If we reach a limit,
abort anyway and let process monitoring restart us.  This should allow
plenty of buffering without infinitely sucking memory and crashing the
device.