init: procd: don't exit, but return.
As the init scripts are sourced and called when procd is building all
it's worldview, if you exit at the wrong time, you can exit the actual
process that is building procd's service view, instead of just your
"own" code.  This was particularly apparent with output-dexma, which is
disabled by default.  Enabling it and "save&apply" or "reload_config"
would correctly issue a service change ubus event, but procd had no
information about the service.  Using a return for control flow instead
of hard exit produces the intended behaviour, of a service _registered_
in procd, but not actually enabled or running.

Fixes GATE-1194