output-senseone: don't disable when ntp loses sync
We want to prevent _starting_ the bridge when we don't have NTP sync,
but if we lose sync due to a network outage, our own clock continues
without any meaningful drift for our purposes, so we want to leave it
running.  On a restart, it won't restart the bridge though, which is
perfectly fine.