output-dexma: ui: handle curl errors better
Between Dexma's lack of CORS and their APIs use of plain text responses
to json POSTS, it's a bit tedious to test this nicely.  In the end,
switch to just parsing the string output from dexma, so we can work
around curl's curiosities with trying to give us a progress meter based
on whether or not we're redirecting output.