output-dexma: coalesce readings and sum/merge based on type
We had cases where we wanted to be grouping/summing per phase energy
counters when the cabinet model indicated a three phase breaker, but we
also have variables that must be reported per phase.  Introduce a
coalesce stage, which not only allows these sorts of options, but also
simplifies and compacts the data sent.

Fixes GATE-1252
Fixes GATE-1239