output-dexma: avoid summation of duplicate hw sources
We automatically summed up matching parameter types for the same "did"
(data point id in dexma terms) to handle the per phase energy counters
being rolled up to a single value for a three phase breaker.  However in
some cases, if a duplicate message was received, the summation would
happen for the same phase, resulting in multiplied values being sent to
dexma.  While the cause of those messages is unidentified, this change
prevents summation of multiple instances, when the source is the same.

Fixes: GATE-1270