output-dexma: enforce a quiet time before processing
We receive NNNN messages when a x minute interval expires, and because
we may need to sum across multi-phase breakers, we want to make sure
we've received the entire interval set before we try and process them.

Failure to do this can mean we only process 2/3 breakers, and then
immediately (next flush loop, 500ms normally) we publish again the
remaining 1/3 breaker.

Fixes GATE-1273