output-dexma: don't re-summ on connection failure
On connection failure, the original unsummed data was pushed back into
the queue, which resulted in resumming.  Properly push the prepared, but
failed to post, data back to the front of the queue, and make sure that
the coalescing functions don't attempt to remodify it.

This manifested as the first reading to be posted after a network
interruption to be a multiple of the correct value, for any value type
that required summing, such as energy from a power bar multi channel
device.

Fixes: GATE-1286