on-charging: pause before restructure config
We're going to need modbus connection information for our devices, as we
need to write to them.  we're not (yet) making any sort of plugin
magical for writing, as it's so arbitrary, and we'd be jsut inventing
apis with no users, even the harbour solution code isn't usefully
relevant here.