io-charging-on: Fix oversubscription behaviour
Instead of jsut reporting the available power to each (required to
allllow the oversusbcription in the first place) make sure to subtract
out what each charger is already using, so that only _excess_ is
oversubscribed, not what is already in use.

Before, with 200A available to chargers, 100 of it in use by charger1,
we were assigning 200A to both, instead of 200 to charger1 and 100 to
charger2.  (ie, each of them are allowed to oversubscribe the
_remaining_ 100A)  New figures correctly handle overload load shedding
down in proportion to useage as well.

Fixes GATE-1336