output-senseone: service+bridge instead of pure bridge.
By using a service, that listens for the interval data and republishes
to a new tree, we can support remapping, along with better process
monitoring, and more finegrained topic remapping.  Downsides are of
course, a new service, and more complicated service monitoring.

This also adds support for configurable security (mostly used for
testing) and statsd.

Fixes GATE-1357